途中経過

+++ src/player-events/audioTrackAdd.ts
@@ -1,14 +1,16 @@
+import { readFileSync } from 'fs';
 import { setTimeout } from 'timers/promises';
 
 import type { GuildQueue, Track } from 'discord-player';
-import type { TextBasedChannel } from 'discord.js';
-import { EmbedBuilder } from 'discord.js';
+import type { TextBasedChannel, User } from 'discord.js';
+import { EmbedBuilder, formatEmoji } from 'discord.js';
 
 import { ExtendedPlayerEvent } from '../interface';
+import type { MusicBot } from '../MusicBot';
 
 export default class extends ExtendedPlayerEvent {
-    public constructor() {
-        super({
+    public constructor(client: MusicBot) {
+        super(client, {
             name: 'audioTrackAdd',
             once: false,
         });
@@ -17,23 +19,37 @@ export default class extends ExtendedPlayerEvent {
     public override execute = async (
         queue: GuildQueue<{
             channel: TextBasedChannel;
+            requestor: User;
         }>,
         track: Track,
     ): Promise<void> => {
-        await queue.metadata.channel
-            .send({
-                embeds: [
-                    new EmbedBuilder()
-                        .setColor('Random')
-                        .setTitle('新しいトラックが追加されました！')
-                        .setDescription(
-                            `[[${track.author}] ${track.title} - (${track.duration})](${track.url})`,
-                        ),
-                ],
-            })
-            .then(async message => {
-                await setTimeout(3000);
-                await message.delete();
-            });
+        const message = await queue.metadata.channel.send({
+            embeds: [
+                new EmbedBuilder()
+                    .setColor('Random')
+                    .setTitle(
+                        `${formatEmoji(
+                            this.client._emojis.owoSaber,
+                            true,
+                        )} 新しいトラックが追加されました！`,
+                    )
+                    .setURL(track.url)
+                    .setDescription(`[${track.author}] ${track.title} - (${track.duration})`),
+            ],
+        });
+
+        const data = JSON.parse(readFileSync('./src/json/data.json', 'utf-8')) as {
+            [key in string]: {
+                auto: boolean;
+                sec: number;
+            };
+        };
+
+        const _req = data[queue.metadata.requestor.id];
+
+        if (_req?.auto) {
+            await setTimeout(_req.sec * 1000);
+            await message.delete();
+        }
     };
 }
